move snapshot tests back
