--- cot/tests/cli.rs
@@ -1,2 +1,0 @@
-#[cfg(not(miri))]
-mod snapshot_testing;
