--- cot/tests/snapshot_testing/mod.rs
@@ -1,1 +1,0 @@
-mod migration;
